@@ -16,6 +16,6 @@
  * @format
  */
 
-module.exports = {
-  presets: [require.resolve('@docusaurus/core/lib/babel/preset')],
+export default {
+  presets: ['@docusaurus/babel/preset'],
 };
